+++ nodes/knxUltimate.js
@@ -49,7 +49,7 @@ module.exports = function (RED) {
             var _dpt = (typeof dpt == "undefined" || dpt == "") ? "" : " DPT" + dpt;
             node.status({ fill: fill, shape: shape, text: _GA + payload + ((node.listenallga && node.server.statusDisplayDeviceNameWhenALL) === true ? " " + _devicename : "") + (node.server.statusDisplayDataPoint === true ? _dpt : "") + (node.server.statusDisplayLastUpdate === true ? " (" + dDate.getDate() + ", " + dDate.toLocaleTimeString() + ")" : "") + " " + text });
             // 16/02/2020 signal errors to the server
-            if (fill.toUpperCase() == "RED") {
+            if (fill.toUpperCase() === "RED") {
                 if (node.server) {
                     var oError = { nodeid: node.id, topic: node.outputtopic, devicename: _devicename, GA: _GA, text: text };
                     node.server.reportToWatchdogCalledByKNXUltimateNode(oError);